Fix Carbon TypeError by casting config values to int or float

+++ src/LogHandler.php
@@ -23,10 +23,10 @@ class LogHandler
     public function handle(MessageLogged $event)
     {
         // Get the log levels specified in the config file
-        $log_types = explode(',',config('log-alarm.log_type'));
+        $log_types = explode(',', config('log-alarm.log_type'));
 
         // Check if the log level of the message is one of the log levels specified in the config file
-        if (in_array($event->level,$log_types)) {
+        if (in_array($event->level, $log_types)) {
 
             // Check if the message contains the specific string specified in the config file
             if ($this->containsSpecificString($event->message)) {
@@ -36,7 +36,7 @@ class LogHandler
             }
         }
     }
-    
+
     /**
      * containsSpecificString
      *
@@ -62,7 +62,7 @@ class LogHandler
         // strpos() returns false.
         return empty($specificString) ? true : strpos($message, $specificString) !== false;
     }
-    
+
     /**
      * logError
      *
@@ -78,21 +78,21 @@ class LogHandler
     protected function logError(MessageLogged $event)
     {
 
-        $log_message =  'LOG_LEVEL: '.$event->level.' | LOG_MESSAGE: '.$event->message;
+        $log_message =  'LOG_LEVEL: ' . $event->level . ' | LOG_MESSAGE: ' . $event->message;
 
         $log_alarm_cache_key_enc = md5($log_message);
-        
+
         // Retrieve the current error logs from the cache or initialize an empty array if no logs exist
         $errorLogs = Cache::get($log_alarm_cache_key_enc, []);
-        
+
         // Add a new log with the current timestamp to the array of error logs
         $errorLogs[] = Carbon::now();
-        
+
         // Get the time in minutes to consider an error log as recent
-        $log_time_frame = config('log-alarm.log_time_frame');
+        $log_time_frame = (float) config('log-alarm.log_time_frame');
 
         // Get specified number of error logs in time frame
-        $log_per_time_frame = config('log-alarm.log_per_time_frame');
+        $log_per_time_frame = (int) config('log-alarm.log_per_time_frame');
 
         // Filter out logs that are older than the specified time frame
         $errorLogs = array_filter($errorLogs, function ($timestamp) use ($log_time_frame) {
@@ -100,21 +100,21 @@ class LogHandler
         });
 
         // Save the updated logs back to the cache with an expiration time of 1 minute
-        Cache::put($log_alarm_cache_key_enc, $errorLogs, Carbon::now()->addMinutes($log_time_frame)); 
+        Cache::put($log_alarm_cache_key_enc, $errorLogs, Carbon::now()->addMinutes($log_time_frame));
 
         // Check if there have been a specified number of error logs in time frame (in the last minute for example)
         if (count($errorLogs) >= $log_per_time_frame) {
-            
+
             // Retrieve the time of the last notification from the cache or initialize null if no notification time exists
-            $last_notification_time = Cache::get($this->notification_cache_key.'_'.$log_alarm_cache_key_enc);
+            $last_notification_time = Cache::get($this->notification_cache_key . '_' . $log_alarm_cache_key_enc);
 
             // Get the delay between notifications from the config file
-            $delay_between_alarms = config('log-alarm.delay_between_alarms');
+            $delay_between_alarms = (float) config('log-alarm.delay_between_alarms');
 
             // Send notification only if last notification was sent more than 5 minutes ago
             // The Carbon library is used to compare the current time with the time of the last notification
             if (!$last_notification_time || Carbon::now()->diffInMinutes($last_notification_time) >= $delay_between_alarms) {
-                
+
                 // Get the message to be sent in the notification from the config file
                 $message = empty(config('log-alarm.notification_message')) ? $log_message : config('log-alarm.notification_message');
 
@@ -125,7 +125,7 @@ class LogHandler
 
                 // Update the time of the last notification in the cache
                 // The notification is set to expire in the delay between alarms specified in the config file
-                Cache::put($this->notification_cache_key.'_'.$log_alarm_cache_key_enc, Carbon::now(), Carbon::now()->addMinutes($delay_between_alarms)); 
+                Cache::put($this->notification_cache_key . '_' . $log_alarm_cache_key_enc, Carbon::now(), Carbon::now()->addMinutes($delay_between_alarms));
             }
         }
     }
